[INSTALLER] Move installer config, commands and controllers to dedicated PimcoreInstallBundle
