+++ pimcore/lib/Pimcore/Bundle/InstallBundle/Resources/config/services.yml
@@ -1,0 +1,21 @@
+services:
+    _defaults:
+        autowire: true
+        autoconfigure: true
+        public: false
+
+    Pimcore\Bundle\InstallBundle\Controller\:
+        resource: '../../Controller'
+        public: true
+        tags: ['controller.service_arguments']
+
+    Pimcore\Bundle\InstallBundle\Command\:
+        resource: '../../Command'
+        public: true
+        tags: ['console.command']
+
+    Pimcore\Composer\PackageInfo: ~
+
+    Pimcore\Install\Installer: ~
+
+    Pimcore\Install\Profile\ProfileLocator: ~
